requirements.txt
python-dotenv~=1.0.0
Scweet~=1.8
requests~=2.31.0
twscrape~=0.7.0
tenacity~=8.2.3
psycopg2~=2.9.7